+++ resources/views/partials/result.blade.php
@@ -1,5 +1,13 @@
-<div class="row b-t">
-    @forelse($result as $item)
+@section('search', $query)
+
+@empty(!$radios)
+    <div class="row b-b v-center pl-3 pr-4">
+        {!! $radios !!}
+    </div>
+@endempty
+
+<div class="row">
+    @forelse($results as $item)
 
         <a href="{{$item->searchUrl()}}" class="block wrapper-sm dropdown-item" style="font-size: 0.82857rem;">
 
@@ -20,15 +28,21 @@
 
     @empty
 
-        <p class="ml-3 mr-3 mb-0 text-center">
-            {{ __('There are no records in this view.') }}
-        </p>
+        <div class="text-center bg-white pt-5 pb-5 w-full">
+            <h3 class="font-thin">
+                <i class="icon-magnifier-remove block m-b"></i>
+                {{ __('Nothing found.') }}
+            </h3>
+
+            {{ __('Try changing the query or type.') }}
+        </div>
+
 
     @endforelse
 
-    @includeWhen($result instanceof \Illuminate\Contracts\Pagination\Paginator && $result->total() > 0,
+    @includeWhen($results instanceof \Illuminate\Contracts\Pagination\Paginator && $results->total() > 0,
         'platform::layouts.pagination',
-        ['paginator' => $result]
+        ['paginator' => $results]
       )
 
 </div>